# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}": {
      "get": {
        "operationId": "PublicUsersController_getPublicProfile",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–50, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "minimum": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Target user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public profile and products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public profile and active listings for a user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by account status",
            "schema": {
              "example": "banned",
              "type": "string",
              "enum": [
                "active",
                "paused",
                "banned"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search by name or phone (1–100 chars)",
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "example": "Ahmed",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (0–10000, default 0)",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersListResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users with optional filters (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user details for moderation page (admin only)",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user (soft-delete, admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/listings": {
      "get": {
        "operationId": "AdminController_listUserListings",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter user listings by status for admin view",
            "schema": {
              "example": "available",
              "type": "string",
              "enum": [
                "available",
                "sold"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (0–10000, default 0)",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated user listings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListingsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List user listings for admin view (read-only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/admins": {
      "get": {
        "operationId": "AdminController_listAdmins",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100, default 20)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (0–10000, default 0)",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of admin users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAdminsListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all admins (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/admins/{id}": {
      "post": {
        "operationId": "AdminController_promoteAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID to promote to admin",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User promoted to admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Promote a user to admin (admin only)",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_demoteAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin user ID to demote",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin demoted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Demote an admin to regular user (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminController_updateUserStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user's status (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/warnings": {
      "post": {
        "operationId": "AdminController_createWarning",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarningDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Warning created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarningResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Target user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue a warning to a user (admin only)",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Place API",
    "description": "Online marketplace REST API — user registration, product listings, real-time chat, ratings, and admin moderation.\n\n> **WebSocket (Chat):** Real-time messaging uses a Socket.io gateway at the `/chat` namespace. That interface is not documented here; see the project README for event names and payloads.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "example": null,
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
...
      "SuccessDataDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/SuccessDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "name": {
            "type": "string",
            "example": "Jana Ahmed"
          },
          "member_since": {
            "type": "string",
            "example": "2025-02-22T10:00:00.000Z"
          },
          "ads_count": {
            "type": "number",
            "example": 10
          },
          "rate": {
            "type": "string",
            "example": "4.50"
          },
          "avatar": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedFileDto"
              }
            ]
          },
          "contactInfo": {
            "type": "object",
            "example": "+201000000001",
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "id",
          "name",
          "member_since",
          "ads_count",
          "rate"
        ]
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "owner": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RelatedUserDto"
              }
            ]
          },
          "category": {
            "type": "string",
            "example": "Electronics"
          },
          "subcategory": {
            "type": "object",
            "example": "Mobile Phones",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Used Laptop"
          },
          "description": {
            "type": "object",
            "example": "Good condition, 1 year old",
            "nullable": true
          },
          "price": {
            "type": "number",
            "example": 1500
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "address_text": {
            "type": "object",
            "example": "10 Tahrir Square",
            "nullable": true
          },
          "details": {
            "type": "object",
            "example": {
              "condition": "used",
              "brand": "Apple",
              "storage": "256GB"
            },
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "available",
            "enum": [
              "available",
              "sold",
              "archived"
            ]
          },
          "is_negotiable": {
            "type": "boolean",
            "example": false
          },
          "preferred_contact_method": {
            "type": "string",
            "example": "both",
            "enum": [
              "phone",
              "chat",
              "both"
            ]
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "seller_rate": {
            "type": "object",
            "example": "4.50",
            "description": "Seller average rating — only present in search results",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          }
        },
        "required": [
          "id",
          "owner",
          "category",
          "name",
          "price",
          "city",
          "status",
          "is_negotiable",
          "preferred_contact_method",
          "created_at",
          "updated_at",
          "images"
        ]
      },
      "PublicUserProfileDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            }
          }
        },
        "required": [
          "user",
          "products"
        ]
      },
      "PublicUserProfileResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/PublicUserProfileDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Ahmed Mohamed"
          },
          "phone": {
            "type": "string",
            "example": "+201012345678"
          },
          "status": {
            "type": "string",
            "example": "active",
            "enum": [
              "active",
              "paused",
              "banned"
            ]
          },
          "is_admin": {
            "type": "boolean",
            "example": true
          },
          "published_products_count": {
            "type": "number",
            "example": 12
          },
          "created_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "status",
          "is_admin",
          "published_products_count",
          "created_at",
          "updated_at"
        ]
      },
      "AdminUsersListDataDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserDto"
            }
          }
        },
        "required": [
          "users"
        ]
      },
      "AdminUsersListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUsersListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "AdminUserDetailsDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AdminUserDetailsDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "AdminUserDetailsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserDetailsDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
      "AdminUserListingsDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserListingDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "AdminUserListingsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserListingsDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "AdminAdminsListDataDto": {
        "type": "object",
        "properties": {
          "admins": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserDto"
            }
          }
        },
        "required": [
          "admins"
        ]
      },
      "AdminAdminsListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminAdminsListDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "AdminUserDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AdminUserDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "AdminUserResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "banned"
            ],
            "description": "New account status for the user",
            "example": "banned"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateWarningDto": {
        "type": "object",
        "properties": {
          "targetUserId": {
            "type": "number",
            "description": "ID of the user to warn",
            "example": 42,
            "minimum": 1
          },
          "message": {
            "type": "string",
            "description": "Warning message (2–2000 chars)",
            "example": "Your listing violated our terms of service.",
            "minLength": 2,
            "maxLength": 2000
          }
        },
        "required": [
          "targetUserId",
          "message"
        ]
      },
...
      "WarningDataDto": {
        "type": "object",
        "properties": {
          "warning": {
            "$ref": "#/components/schemas/WarningDto"
          }
        },
        "required": [
          "warning"
        ]
      },
      "WarningResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/WarningDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}